+++ core/spoofax.eclipse/src/main/java/mb/spoofax/eclipse/job/FakeMultiRule.java
@@ -1,0 +1,11 @@
+package mb.spoofax.eclipse.job;
+
+import org.eclipse.core.runtime.jobs.ISchedulingRule;
+import org.eclipse.core.runtime.jobs.MultiRule;
+
+public class FakeMultiRule extends MultiRule {
+    // HACK: extend MultiRule with empty list to make users of this rule compatible with Resource rules.
+    public FakeMultiRule() {
+        super(new ISchedulingRule[0]);
+    }
+}