+++ example/tiger/manual/tiger.eclipse/src/main/java/mb/tiger/eclipse/TigerEclipseComponent.java
@@ -25,4 +25,6 @@ import mb.tiger.spoofax.TigerScope;
 )
 public interface TigerEclipseComponent extends EclipseLanguageComponent, TigerComponent {
     TigerEditorTracker getEditorTracker();
+
+    @Override TigerCheckCallback.Factory getCheckCallbackFactory();
 }